+++ src/components/floating/primitives/useFloatingNode.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useFloatingNodeId, useFloatingParentNodeId } from "@floating-ui/react";
 
 /**